Cleanup

+++ context/scripts/helpers.sh
@@ -14,9 +14,7 @@ cache::download(){
   local url="$3"
 
   mkdir -p "$CACHE_ROOT/$arch"
-  if [ ! -f "$CACHE_ROOT/$arch/$name" ]; then
-    curl --proto '=https' --tlsv1.2 -sSfL --compressed -o "$CACHE_ROOT/$arch/$name" "$url" >/dev/null 2>&1
-  fi
+  [ -f "$CACHE_ROOT/$arch/$name" ] || curl --proto '=https' --tlsv1.2 -sSfL --compressed -o "$CACHE_ROOT/$arch/$name" "$url" >/dev/null 2>&1
 }
 
 cache::path(){
@@ -144,6 +142,7 @@ version::latest::patch(){
   local next_patch
 
   next_patch=$((patch + 1))
+  echo >&2 curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$minor" "$next_patch")"
   while curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$minor" "$next_patch")" 2>&1 | grep -qE "HTTP/[0-9. ]+ 200"; do
   #while [ "$(curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$minor" "$next_patch")" 2>&1 | grep -E "HTTP/[0-9.]+ [0-9]{3}" | tail -1 | sed -E 's/.* ([0-9]{3}).*/\1/')" != "404" ]; do
     candidate_patch="$next_patch"
@@ -152,6 +151,7 @@ version::latest::patch(){
     curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$minor" "$next_patch")" 2>&1 | grep -E "HTTP/[0-9.]+" 1>&2
     sleep 1
   done
+
   printf "%s.%s.%s" "$major" "$minor" "$candidate_patch"
   if [ "$candidate_patch" != "$patch" ];then
     return 1
@@ -173,6 +173,7 @@ version::latest::minor(){
 
   next_minor=$((minor + 1))
 
+  echo >&2 curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$next_minor" "$patch")"
   while curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$next_minor" "$patch")" 2>&1 | grep -qE "HTTP/[0-9. ]+ 200"; do
 #  while [ "$(curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$next_minor" "$patch")" 2>&1 | grep -E "HTTP/[0-9.]+ [0-9]{3}" | tail -1 | sed -E 's/.* ([0-9]{3}).*/\1/')" != "404" ]; do
     candidate_minor=${next_minor}
@@ -181,6 +182,7 @@ version::latest::minor(){
     curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$major" "$next_minor" "$patch")" 2>&1 | grep -E "HTTP/[0-9.]" 1>&2
     sleep 1
   done
+
   printf "%s.%s.0" "$major" "$candidate_minor"
   if [ "$candidate_minor" != "$minor" ];then
     return 1
@@ -203,9 +205,13 @@ version::latest::major(){
   local next_major
 
   next_major=$((major + increment))
+  echo >&2 curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$next_major" "$minor" "$patch")"
   while curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$next_major" "$minor" "$patch")" 2>&1 | grep -qE "HTTP/[0-9. ]+ 200"; do
     candidate_major=${next_major}
     next_major=$((next_major + increment))
+    echo >&2 curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$next_major" "$minor" "$patch")"
+    curl --proto '=https' --tlsv1.2 -L -I -o /dev/null -v "$("$urlfunction" "$platform" "$next_major" "$minor" "$patch")" 2>&1 | grep -E "HTTP/[0-9.]+" 1>&2
+    sleep 1
   done
   printf "%s.0.0" "$candidate_major"
   if [ "$candidate_major" != "$major" ];then
